fixed warning

+++ LoopieEditor/src/Interfaces/Editor/EditorMenuInterface.cpp
@@ -192,9 +192,9 @@ namespace Loopie {
 
 			char title[25];
 			sprintf_s(title, 25, "Framerate %.1f", m_fpsLog.back());
-			ImGui::PlotHistogram("##framerate", &m_fpsLog[0], m_fpsLog.size(), 0, title, 0.0f, 100.0f, ImVec2(310, 100));
+			ImGui::PlotHistogram("##framerate", &m_fpsLog[0], (int)m_fpsLog.size(), 0, title, 0.0f, 100.0f, ImVec2(310, 100));
 			sprintf_s(title, 25, "Milliseconds %.1f", m_msLog.back());
-			ImGui::PlotHistogram("##milliseconds", &m_msLog[0], m_msLog.size(), 0, title, 0.0f, 40.0f, ImVec2(310, 100));
+			ImGui::PlotHistogram("##milliseconds", &m_msLog[0], (int)m_msLog.size(), 0, title, 0.0f, 40.0f, ImVec2(310, 100));
 
 		}
